backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
Pillow